fix: ajout de test Id pour les événements dans Activités
