+++ app/src/types/notification-list-item.ts
@@ -1,0 +1,8 @@
+export enum NotificationTypeEnum {
+  BasicMessage = 'BasicMessage',
+  CredentialOffer = 'Offer',
+  ProofRequest = 'ProofRecord',
+  Revocation = 'Revocation',
+  Custom = 'Custom',
+  Proof = 'Proof',
+}
